Split logging out to separate file
